wip: testing city menu grouping
