+++ pages/index.jsx
@@ -5,12 +5,14 @@ import CitySelector from '@/components/home/CitySelector'
 import Block from '@/components/layout/Block'
 import {
   getCommonApiContent,
-  getHeroFromNode,
   resolvePath,
   getLandingPageQueryParams,
   getThemeHeroImages,
   NOT_FOUND,
 } from '@/lib/ssr-api'
+
+import { getHeroFromNode } from '@/lib/ssr-helpers'
+
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import getConfig from 'next/config'
 import { getResource } from 'next-drupal'
@@ -76,9 +78,15 @@ export async function getStaticProps(context) {
   }
 }
 
-const HomePage = ({ menu, footerMenu, node, themes, municipalities }) => {
-  useHydratePage({ node, municipalities, footerMenu, menu })
-
+const HomePage = ({
+  menu,
+  footerMenu,
+  node,
+  themes,
+  citiesMenu,
+  municipalities,
+}) => {
+  useHydratePage({ node, municipalities, footerMenu, menu, citiesMenu })
   const hero = getHeroFromNode(node)
   const { field_description, field_content, title } = node
