+++ src/components/cities/LocalInformation.jsx
@@ -13,7 +13,7 @@ import { IconAngleDown } from '@/components/Icons'
 import { useTranslation } from 'next-i18next'
 import { CSSTransition } from 'react-transition-group'
 import useSWR from 'swr'
-import { getLinks } from '@/lib/ssr-api'
+import { getLinks } from '@/lib/ssr-helpers'
 import TextLink from '../TextLink'
 import { DotsLoader } from '../Loaders'
 import { IconExclamationCircle } from '../Icons'
